Update Colors And Dark Mode Toggle to Improve A11Y
Updated the colors used on the website to increase
the contrast between text and background to meet
WCAG 2 AA contrast thresholds.

Also updated the dark mode toggle to not use a label
as a wrapper element. Further improved accessibility
of this control by labeling the input itself.

And finally, added page titles to controllers that
were missing them.