Fix Dark Mode Flicker and Handle Invalid Locales (#89)
* Fix Dark Mode Flicker and Handle Invalid Locales

Updated the dark mode logic to add a blocking js
script until the theme is set to avoid a brief
flicker of light mode on page load (when dark
mode is enabled).

Also updated the set locale plug to check the
given locale against the list of valid locales
and redirect to the home page if the locale
is not valid.

Finally, updated a helper function used by the
locale menu to redirect the user to the
expected locale when they click on one.

* Update current_page_locale_path Helper Function

Updated the current_page_locale helper function to
only replace the locale if it appears at the beginning
request path. This prevents non-localized paths from
being malformed if they contain a locale name in
the slug.

Also added a couple more tests for this function.