Fix Review App URL Configuration (#133)
Updated the deploy process to generate the xml
and feed files when the app is deployed so that
the correct URL is used for them. This is helpful
for review apps which previously had sitemap links
pointing to "beta.elixirschool.com".

This also resolves an issue where live view would
not work on review apps since the host url
configured with `Endpoint` was always
"beta.elixirschool.com" but the request origin
was from the review app domain.